sign of position

+++ src/lib.rs
@@ -458,7 +458,7 @@ impl<SPI, CS, EN, E> Tmc5160<SPI, CS, EN>
 
     /// get the current position
     pub fn get_position(&mut self) -> Result<f32, Error<E>> {
-        self.read_register(Registers::XACTUAL).map(|val| val.data as f32 / self._step_count / 400.0)
+        self.read_register(Registers::XACTUAL).map(|val| (val.data as i32) as f32 / self._step_count / 400.0)
     }
 
     /// set the current position
